requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.25
SQLAlchemy>=2.0
alembic==1.16.5
psycopg[binary]==3.2.10
python-dotenv==1.1.1
